♻️ Improve: `client::leaderboard` modlue structure [#63]
- ✏️ Rename: to `user_leaderboard`
- 🚚 Move: to `param` module
- ✏️ Rename: `LeaderboardSearchCriteria` to `SearchCriteria`
